backend/requirements.txt
alembic==1.18.4
blinker==1.9.0
click==8.3.1
exceptiongroup==1.3.1
Flask==3.0.3
flask-cors==6.0.2
Flask-JWT-Extended==4.6.0
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
greenlet==3.3.2
google-generativeai==0.8.5
iniconfig==2.3.0
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.10
MarkupSafe==3.0.3
packaging==26.0
pluggy==1.6.0
psycopg2-binary==2.9.10
PyJWT==2.12.1
pytest==8.2.0
python-dotenv==1.0.1
razorpay==1.4.1
SQLAlchemy==2.0.48
tomli==2.4.1
typing_extensions==4.15.0
Werkzeug==3.1.7
